Create skbio_alt module and move pcoa into it
