+++ sucpp/Makefile
@@ -83,13 +83,13 @@ endif
 
 CPPFLAGS += -Wall  -std=c++11 -pedantic -I. $(OPT) -fPIC -L$(CONDA_PREFIX)/lib
 
-test: tree.o test_su.cpp biom.o unifrac.o api.o
-	$(CXX) $(CPPFLAGS) $(EXEFLAGS) test_su.cpp -o test_su tree.o biom.o unifrac.o api.o -llz4 -lcblas -lpthread
-	$(CXX) $(CPPFLAGS) $(EXEFLAGS) test_api.cpp -o test_api tree.o biom.o unifrac.o api.o -llz4 -lcblas -lpthread
+test: tree.o test_su.cpp biom.o unifrac.o skbio_alt.o api.o
+	$(CXX) $(CPPFLAGS) $(EXEFLAGS) test_su.cpp -o test_su tree.o biom.o unifrac.o skbio_alt.o api.o -llz4 -lcblas -lpthread
+	$(CXX) $(CPPFLAGS) $(EXEFLAGS) test_api.cpp -o test_api tree.o biom.o unifrac.o skbio_alt.o api.o -llz4 -lcblas -lpthread
 
-main: tree.o biom.o unifrac.o cmd.o api.o
-	$(CXX) $(CPPFLAGS) $(EXEFLAGS) su.cpp -o ssu tree.o biom.o unifrac.o cmd.o api.o -lhdf5_cpp -llz4 -lcblas -lpthread
-	$(CXX) $(CPPFLAGS) $(EXEFLAGS) faithpd.cpp -o faithpd tree.o biom.o unifrac.o cmd.o api.o -lhdf5_cpp -llz4 -lcblas -lpthread
+main: tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o
+	$(CXX) $(CPPFLAGS) $(EXEFLAGS) su.cpp -o ssu tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o -lhdf5_cpp -llz4 -lcblas -lpthread
+	$(CXX) $(CPPFLAGS) $(EXEFLAGS) faithpd.cpp -o faithpd tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o -lhdf5_cpp -llz4 -lcblas -lpthread
 	cp ssu ${PREFIX}/bin/
 	cp faithpd ${PREFIX}/bin/
 
@@ -107,14 +107,16 @@ rapi_test: main
 	echo LDFLAGS=-llz4 -lcblas >> ~/.R/Makevars
 	Rscript R_interface/rapi_test.R
 	
-api: tree.o biom.o unifrac.o cmd.o 
-	$(CXX) $(CPPFLAGS) api.cpp -c -o api.o -fPIC
-	$(CXX) $(LDDFLAGS) -o libssu.so tree.o biom.o unifrac.o cmd.o api.o -lc -lhdf5_cpp -llz4 -lcblas -L$(PREFIX)/lib
+api: tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o
+	$(CXX) $(LDDFLAGS) -o libssu.so tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o -lc -lhdf5_cpp -llz4 -lcblas -L$(PREFIX)/lib
 	cp libssu.so ${PREFIX}/lib/
 
 capi_test: api
 	gcc -std=c99 capi_test.c -lssu -L${PREFIX}/lib -Wl,-rpath,${PREFIX}/lib -o capi_test
 	export LD_LIBRARY_PATH="${PREFIX}/lib":"./capi_test"
+
+api.o: api.cpp api.hpp unifrac.hpp skbio_alt.hpp biom.hpp tree.hpp
+	$(CXX) $(CPPFLAGS) api.cpp -c -o api.o -fPIC
 
 %.o: %.cpp %.hpp
 	$(CXX) $(CPPFLAGS) -c $< -o $@
